+++ orttraining/orttraining/python/training/__init__.py
@@ -1,12 +1,11 @@
-#-------------------------------------------------------------------------
+# -------------------------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
-#--------------------------------------------------------------------------
+# --------------------------------------------------------------------------
 from onnxruntime.capi._pybind_state import TrainingParameters
 from onnxruntime.capi.training.training_session import TrainingSession
 
 
 from .orttrainer_options import ORTTrainerOptions
-from . import model_desc_validation
-from . import amp
 from .orttrainer import TrainStepInfo
+from . import amp, optim, model_desc_validation
